Gavin's Update to Statistical Methods
Restructured the ending part of the document so that the Analysis comes after their respective set of Figures. (i.e. put Figures 5-8 after the analysis for Figures 1-4) Added the Linear Models to the graphs. Included a third graph where, instead of excluding Africa, we exclude countries of Life expectancy below 65. Didn't create any better of an R^2 value than the original graph but it was worth a shot. Finished the analysis for the Statistical Methods graphs.
Also changed the read_cvs at the beginning of the document to call cia.cvs from this github page rather than call a local directory. This way we can all use the exact same code rather than changing the directory for our respective R clients.